requirements.txt
paho-mqtt==2.1.0
influxdb-client==1.50.0
pyyaml>=6.0
streamlit>=1.56.0
pandas>=2.0.0
scikit-learn>=1.8.0
plotly>=5.0.0
numpy>=1.24.0